# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UniHub Workshop API
  version: 1.0.0
  description: "REST API specification for the **UniHub Workshop** system — university\
    \ workshop\nregistration & check-in platform. This document is the authoritative\
    \ contract\nbetween backend (NestJS Modular Monolith) and frontend (Next.js Web\
    \ Portal +\nExpo Mobile App).\n\n## Conventions\n\n- **Versioning:** URL-based,\
    \ prefix `/api/v1`.\n- **Authentication:** JWT Bearer (15-minute access token)\
    \ + refresh token (7 days).\n  Refresh token transport is **client-type specific**\
    \ (see `authentication.md` INV-02):\n  - **Web client:** `refresh_token` is set\
    \ as an `HttpOnly; SameSite=Strict` cookie\n    on `/auth/login` and `/auth/refresh`\
    \ responses. JavaScript cannot read it.\n    Access token is returned in the JSON\
    \ body and kept **in memory only** (not localStorage).\n  - **Mobile client:**\
    \ `refresh_token` is also included in the JSON response body\n    (in addition\
    \ to the cookie) because mobile has no cookie jar — app stores it in\n    Android\
    \ Keystore / iOS Keychain.\n- **Idempotency:** `POST /registrations` and `POST\
    \ /payments` require an\n  `X-Idempotency-Key: <UUID v4>` **HTTP header** (client-generated,\
    \ reuse on retry).\n  Server uses 3-state lifecycle (`in_progress` / `completed`\
    \ / `unresolved`).\n  See ADR-08, ADR-15.\n- **Optimistic Locking:** Admin updates\
...
    \ `VND`).\n"
  contact:
...
  license:
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /workshops/{workshopId}:
    get:
      tags:
      - Catalog
      summary: Get workshop detail
      description: "Full workshop detail page including speaker bio, room floor plan,

        and AI summary (if `summary_status='DONE'`).

        "
      operationId: getWorkshop
      security:
      - {}
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/WorkshopIdPath'
      responses:
        '200':
          description: Workshop detail
          headers:
            Cache-Control:
              schema:
                type: string
                default: public, max-age=10
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/workshops:
    get:
      tags:
      - Catalog Admin
      summary: List all workshops (admin view)
      description: 'Includes all statuses (`draft`, `open`, `closed`, `cancelled`).

        **Role:** `btc`.

        '
      operationId: adminListWorkshops
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - DRAFT
          - OPEN
          - COMPLETED
          - CANCELLED
      - name: q
        in: query
        description: Search by title (ILIKE)
        schema:
          type: string
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Admin workshop list
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkshopAdmin'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Catalog Admin
      summary: Create a workshop (default `status=draft`)
      operationId: adminCreateWorkshop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkshopCreateRequest'
      responses:
        '201':
          description: Workshop created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopAdmin'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 'Room schedule conflict (`workshop.room_conflict`) — another

            workshop already booked in this room and time range.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/workshops/{workshopId}:
    get:
      tags:
      - Catalog Admin
      summary: Get workshop detail (admin)
      operationId: adminGetWorkshop
      parameters:
      - $ref: '#/components/parameters/WorkshopIdPath'
      responses:
        '200':
          description: Admin workshop detail (includes `version`)
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopAdmin'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Catalog Admin
      summary: Update workshop (with Optimistic Locking)
      description: '**Concurrency:** This endpoint requires `If-Match` carrying the

        current `version`. If the underlying row has been updated since the

        client last read it, returns `412 Precondition Failed`.


        **Side effect:** Changing `roomId`, `startsAt`, or `endsAt` on a

        workshop with active registrations triggers `workshop.changed`

        notifications to all registered students.


        **Edge case:** Lowering `seatsTotal` below `(seatsTotal - seatsAvailable)`

        (i.e., below the count of confirmed registrations) returns 422.

        '
      operationId: adminPatchWorkshop
      parameters:
      - $ref: '#/components/parameters/WorkshopIdPath'
      - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkshopPatchRequest'
      responses:
        '200':
          description: Workshop updated (`version` incremented)
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopAdmin'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          description: 'Validation error — e.g., `workshop.seats_total_below_registered`,

            `workshop.invalid_time_range`.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/workshops/{workshopId}/publish:
    post:
      tags:
      - Catalog Admin
      summary: Publish a draft workshop (`draft` → `open`)
      description: 'Requires `roomId` and `speakerId` to be set. Validates the room
        is

        not double-booked.

        '
      operationId: adminPublishWorkshop
      parameters:
      - $ref: '#/components/parameters/WorkshopIdPath'
      - $ref: '#/components/parameters/IfMatchHeader'
      responses:
        '200':
          description: Workshop published
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopAdmin'
        '409':
          description: '`workshop.cannot_publish` — already open, or missing `roomId`/`speakerId`.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /admin/workshops/{workshopId}/emergency-update:
    patch:
      tags:
      - Catalog Admin
      summary: Emergency update a published workshop
      description: 'Updates scheduling fields (`roomId`, `startsAt`, `endsAt`) of an
        already published workshop without changing its status.
        Re-validates room time conflicts if scheduling fields are modified.

        **Concurrency:** Requires `If-Match` header with the current `version`.
        Mismatch returns `412 Precondition Failed`.

        **Edge case:** Lowering `seatsTotal` is NOT allowed through this endpoint
        — use the regular PATCH on draft workshops instead.
        '
      operationId: adminEmergencyUpdateWorkshop
      parameters:
      - $ref: '#/components/parameters/WorkshopIdPath'
      - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roomId:
                  type: string
                  format: uuid
                  description: New room UUID. Null to remove assignment.
                startsAt:
                  type: string
                  format: date-time
                  description: New start time
                endsAt:
                  type: string
                  format: date-time
                  description: New end time
              minProperties: 1
              description: At least one field must be provided.
      responses:
        '200':
          description: Workshop updated successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopAdmin'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: '`workshop.room_conflict` — the new room/time overlaps another
            workshop.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /admin/workshops/{workshopId}/cancel:
    post:
      tags:
      - Catalog Admin
      summary: Cancel a workshop
      description: '**Cascading effects (in one transaction):**

        - All `paid`/`confirmed`/`pending` registrations → `cancelled`.

        - For each `paid` registration → enqueue refund job.

        - For all registrations → enqueue `workshop_cancelled` notification.

        '
      operationId: adminCancelWorkshop
      parameters:
      - $ref: '#/components/parameters/WorkshopIdPath'
      - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - reason
              properties:
                reason:
                  type: string
                  minLength: 5
                  maxLength: 500
                notifyRegistered:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Workshop cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopAdmin'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /admin/workshops/{workshopId}/stats:
    get:
      tags:
      - Catalog Admin
      summary: Aggregated workshop stats
      description: 'Cross-module aggregation: registrations (booking) + revenue (payment)
        +

        check-in attendance (checkin). Heavily cached (5 min).

        '
      operationId: adminGetWorkshopStats
      parameters:
      - $ref: '#/components/parameters/WorkshopIdPath'
      responses:
        '200':
          description: Stats payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopStats'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Short-lived access token (15 min). RS256 signed. Decoded claims:

        `sub` (studentId or staff UUID), `role`, `user_type`,

        `allowedWorkshopIds` (checkin_staff only), `iat`, `exp`.

        Verified by public key — no DB lookup at Layer ① (authentication.md INV-01).

        '
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: HMAC-SHA256 signature of the request body. Per-gateway secret.
  headers:
...
    ETag:
      description: Resource version (the `version` column for OL-protected resources)
      schema:
        type: string
        example: '"42"'
  parameters:
    WorkshopIdPath:
      name: workshopId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatchHeader:
      name: If-Match
      in: header
      required: true
      description: 'Quoted version string from the previous `ETag` response header.

        Used for Optimistic Locking on admin updates.

        '
      schema:
        type: string
        example: '"42"'
    CursorParam:
      name: cursor
      in: query
      description: Opaque cursor from the previous response's `pagination.next_cursor`
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Forbidden:
      description: Authenticated but lacks permission for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found, or hidden by row-level filter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    PreconditionFailed:
      description: '`If-Match` does not match the current resource version. The client

        should re-fetch the resource to get the latest version and re-apply

        the change.

        '
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
      - success
      - error
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
          description: Always false for error responses
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              description: Stable error code clients can switch on
              example: registration.workshop_full
            message:
              type: string
              description: Human-readable error message safe for client display
              example: Workshop đã đầy chỗ
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
              description: Field-level validation details when applicable
        meta:
          $ref: '#/components/schemas/RequestMeta'
...
    RequestMeta:
      type: object
      description: Metadata attached to every API response for tracing
      required:
      - requestId
      - timestamp
      - apiVersion
      properties:
        requestId:
          type: string
          description: Correlation ID for request tracing
          example: req-abc123
        timestamp:
          type: string
          format: date-time
          description: Response timestamp (RFC 3339)
        apiVersion:
          type: string
          description: API version tag
          example: v1
        processingMs:
          type: integer
          description: Server-side processing duration in milliseconds
          example: 42
    FieldError:
      type: object
      description: Field-level validation error details
      required:
      - field
      - rule
      - message
      properties:
        field:
          type: string
          description: Field path that failed validation
          example: email
        rule:
          type: string
          description: Validation rule identifier
          example: isEmail
        message:
          type: string
          description: Human-readable validation message
          example: Invalid email format
        received:
          description: Raw input value (when safe to expose)
    PaginationMeta:
      type: object
      required:
      - limit
      - hasMore
      properties:
        limit:
          type: integer
        nextCursor:
          type: string
          nullable: true
          description: Pass back as `?cursor=` to fetch the next page
        hasMore:
          type: boolean
        total:
          type: integer
          nullable: true
          description: Only populated for offset-aware admin endpoints
...
    WorkshopListItem:
      type: object
      required:
      - id
      - title
      - startsAt
      - endsAt
      - seatsTotal
      - seatsAvailable
      - price
      - currency
      - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        seatsTotal:
          type: integer
          minimum: 1
        seatsAvailable:
          type: integer
          minimum: 0
          description: Cache hint, may be up to 10s stale
        price:
          type: number
          minimum: 0
        currency:
          type: string
          default: VND
        status:
          type: string
          enum:
          - DRAFT
          - OPEN
          - COMPLETED
          - CANCELLED
        speaker:
          allOf:
          - $ref: '#/components/schemas/SpeakerSummary'
          nullable: true
        room:
          allOf:
          - $ref: '#/components/schemas/RoomSummary'
          nullable: true
        isRegistered:
          type: boolean
          nullable: true
          description: True if the authenticated student has an active registration.
            `null` if anonymous.
    WorkshopDetail:
      allOf:
      - $ref: '#/components/schemas/WorkshopListItem'
      - type: object
        properties:
          description:
            type: string
            nullable: true
          speaker:
            $ref: '#/components/schemas/Speaker'
          room:
            $ref: '#/components/schemas/Room'
          summary:
            $ref: '#/components/schemas/AiSummary'
          myRegistrationId:
            type: string
            format: uuid
            nullable: true
    WorkshopAdmin:
      allOf:
      - $ref: '#/components/schemas/WorkshopDetail'
      - type: object
        required:
        - version
        - createdAt
        - updatedAt
        properties:
          version:
            type: integer
            format: int64
            description: Optimistic lock counter — also exposed as `ETag`
          pdfUrl:
            type: string
            nullable: true
          createdBy:
            type: string
            format: uuid
            nullable: true
          createdAt:
            type: string
            format: date-time
          updatedAt:
            type: string
            format: date-time
...
    WorkshopCreateRequest:
      type: object
      required:
      - title
      - startsAt
      - endsAt
      - seatsTotal
      - price
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        speakerId:
          type: string
          format: uuid
          nullable: true
        roomId:
          type: string
          format: uuid
          nullable: true
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        seatsTotal:
          type: integer
          minimum: 1
          maximum: 1000
        price:
          type: number
          minimum: 0
        status:
          type: string
          enum:
          - DRAFT
          - OPEN
          default: DRAFT
    WorkshopPatchRequest:
      type: object
      description: All fields optional (PATCH semantics)
      properties:
        title:
          type: string
        description:
          type: string
        speakerId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        seatsTotal:
          type: integer
          minimum: 1
        price:
          type: number
          minimum: 0
    WorkshopStats:
      type: object
      properties:
        registrations:
          type: object
          properties:
            total:
              type: integer
            byStatus:
              type: object
              additionalProperties:
                type: integer
              example:
                paid: 45
                pending: 3
                confirmed: 10
                cancelled: 2
        checkins:
          type: object
          properties:
            total:
              type: integer
            rate:
              type: number
              minimum: 0
              maximum: 1
        revenue:
          type: object
          properties:
            amount:
              type: number
            currency:
              type: string
              default: VND
    RoomSummary:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        building:
          type: string
          nullable: true
        floor:
          type: integer
          nullable: true
        floorPlanUrl:
          type: string
          nullable: true
    Room:
      allOf:
      - $ref: '#/components/schemas/RoomSummary'
      - type: object
        properties:
          capacity:
            type: integer
          facilities:
            type: object
            additionalProperties: true
            example:
              projector: true
              ac: true
              mic: 2
          createdAt:
            type: string
            format: date-time
...
    SpeakerSummary:
      type: object
      required:
      - id
      - fullName
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        title:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
    Speaker:
      allOf:
      - $ref: '#/components/schemas/SpeakerSummary'
      - type: object
        properties:
          bio:
            type: string
            nullable: true
...
    AiSummary:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - NONE
          - QUEUED
          - PROCESSING
          - DONE
          - FAILED
        text:
          type: string
          nullable: true
          description: Present only when `status=done`
        updatedAt:
          type: string
          format: date-time
          nullable: true
        errorDetail:
          type: string
          nullable: true
          description: Present only when `status=failed`
...